[fix] 파일 구조변경
